Fixed nil values from phrase generation
I was using the wrong name for the line number variable in the get_line function, which is why it would reach the end of the loop and return nil. Also added a phrases dictionary so you get a full phrase from the bot and not just a random "jerk off" euphemism.