Added basic help function and precursor to roll function
This change is so that I can test the dice roll functionality quicker, as the framework for calling the function from Discord is already in place.  Eventually I will also add a real help function that is formatted nicely, but the current menu will have to do.  I would eventually like to add a default instruction for ./roll so that it tells users how to use the command.