slightly more advanced dice roll placeholder
I added a check to see if only ./roll was issued, if so then it gives instructions on how to use the command.  Otherwise it still states that the command is not implemented yet.  Also, I wrote a first version of the function to actually take in a die roll (e.g. 1d20) and turn it into a table of values that will be returned.